ecommerce_api/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
parts==2.1.0
pillow==11.0.0
pipenv==2024.0.2
platformdirs==4.3.6
products==2.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pypyodbc==1.3.6
python-telegram-bot==21.4
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.5
whitenoise==6.8.2
